Move test to Pool.spec
