+++ tests/e2e/specs/Modeler.spec.js
@@ -35,26 +35,6 @@ describe('Modeler', () => {
     cy.get('.modeler').children().should('have.length', emptyChildrenCount + nodeTypes.length);
   });
 
-  it('Can add top and bottom lane', () => {
-    const poolPosition = { x: 200, y: 200 };
-
-    dragFromSourceToDest(
-      'processmaker-modeler-pool',
-      '.paper-container',
-      poolPosition,
-    );
-
-    waitToRenderAllShapes();
-
-    getElementAtPosition(poolPosition)
-      .click()
-      .then($pool => {
-        getCrownButtonForElement($pool, 'lane-above-button').click();
-        getCrownButtonForElement($pool, 'lane-below-button').click();
-      });
-
-  });
-
   it('Update task name', () => {
     const testString = 'testing';
 
@@ -96,20 +76,6 @@ describe('Modeler', () => {
     cy.get('.joint-viewport').find('.joint-type-standard-polyline').click({force: true});
     typeIntoTextInput('[name=\'text\']', testString);
     cy.get('[name=\'text\']').should('have.value', testString);
-  });
-
-  it('Update pool name', () => {
-    const testString = 'testing';
-
-    dragFromSourceToDest(
-      'processmaker-modeler-pool',
-      '.paper-container',
-      200, 200
-    );
-
-    cy.get('.joint-viewport').find('.joint-type-processmaker-modeler-bpmn-pool').click({force: true});
-    typeIntoTextInput('[name=\'name\']', testString);
-    cy.get('[name=\'name\']').should('have.value', testString);
   });
 
   it('Create a simple process', () => {
